split client and server side logic
