+++ gradle.properties
@@ -1,15 +1,18 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
-# Fabric Properties
-# check these on https://modmuss50.me/fabric.html
-minecraft_version=1.16.5
-loader_version=0.14.8
-yarn_mappings=1.16.5+build.10
+org.gradle.parallel=true
 
-fabric_version=0.42.0+1.16
+# Fabric Properties
+# check these on https://fabricmc.net/develop
+minecraft_version=1.19.3
+yarn_mappings=1.19.3+build.1
+loader_version=0.14.14
+
+#Fabric api
+fabric_version=0.73.2+1.19.3
 
 # Mod Properties
 mod_version=0.0.1
-maven_group=com.cce
+maven_group=com.yukii
 archives_base_name=anvildelimited
