+++ src/main/java/com/yukii/anvildelimited/AnvilDelimited.java
@@ -1,5 +1,7 @@
 package com.yukii.anvildelimited;
 
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
@@ -14,7 +16,7 @@ import net.minecraft.world.GameRules;
 
 public class AnvilDelimited implements ModInitializer {
     public static int gameRuleValue;
-    private static final Identifier ID = new Identifier("anvil-delimited", "send-gamerule-value");
+    public static final Identifier ID = new Identifier("anvil-delimited", "send-gamerule-value");
 
     /**
      * Creates the actual gamerule. The lambda will run every time the gamerule value updates,
@@ -35,28 +37,16 @@ public class AnvilDelimited implements ModInitializer {
     public void onInitialize() {
         gameRuleValue = 40;
         registerServerPlayerListener();
-        registerGlobalValueReceiver();
     }
 
-    /**
-     * Registers a global receiver for the purposes of syncing the client and server sides together.
-     */
-    private void registerGlobalValueReceiver() {
-        ClientPlayNetworking.registerGlobalReceiver(AnvilDelimited.ID, (client, handler, buf, responseSender) -> {
-            int i = buf.readInt();
-            gameRuleValue = i < 2 || i > Short.MAX_VALUE ? 40 : i;
-        });
-    }
 
     /**
      * Registers a listener. The lambda executes everytime a player joins a world or a server and sends the player the gamerule's value integer as a packet.
      */
-
     private void registerServerPlayerListener() {
         ServerEntityEvents.ENTITY_LOAD.register(
                 ((entity, world) -> {
-                    if (entity instanceof ServerPlayerEntity) {
-                        ServerPlayerEntity player = (ServerPlayerEntity) entity;
+                    if (entity instanceof ServerPlayerEntity player) {
                         PacketByteBuf sBuf = PacketByteBufs.create();
                         sBuf.writeInt(world.getGameRules().getInt(AnvilDelimited.ANVIL_XP_LIMIT));
                         if (ServerPlayNetworking.canSend(player, ID)) ServerPlayNetworking.send(player, ID, sBuf);
