+++ src/main/resources/anvil-delimited.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.cce.anvildelimited.mixin",
+  "package": "com.yukii.anvildelimited.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "AnvilScreenHandlerMixin"
